# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Median Code API
  description: |
    Production-ready FastAPI code generation and API entity management with Clerk JWT authentication.

    ## Authentication

    All endpoints require Bearer JWT authentication via Clerk. Include your Clerk JWT token in the Authorization header:

    ```
    Authorization: Bearer <your-clerk-jwt-token>
    ```

    ## API Overview

    This API provides two main categories of endpoints:

    ### 1. Entity Management (CRUD Operations)

    Manage your API design entities with full CRUD operations:

    - **Namespaces** (`/namespaces`) - Create and manage namespaces for organizing entities
    - **APIs** (`/apis`) - Create and manage API definitions that group endpoints and tags
    - **Types** (`/types`) - Retrieve primitive and abstract type definitions (read-only)
    - **Validators** (`/validators`) - Retrieve validator definitions (read-only)
    - **Fields** (`/fields`) - Create, read, update, and delete field definitions
    - **Objects** (`/objects`) - Create, read, update, and delete object definitions
    - **Tags** (`/tags`) - Create, read, update, and delete endpoint tags
    - **Endpoints** (`/endpoints`) - Create, read, update, and delete API endpoint definitions

    All list endpoints support filtering by `namespaceId` query parameter.

    ### 2. Code Generation

...

    #### Generated Application Structure

    The generated zip file contains:
...
    - `requirements.txt` - Python dependencies
...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /tags:
    get:
      tags:
        - Tags
      summary: List all tags
      description: Retrieve all endpoint tags. Tags can be filtered by namespace.
      operationId: listTags
      parameters:
        - name: namespaceId
          in: query
          required: false
          schema:
            type: string
          description: Filter tags by namespace ID
      responses:
        '200':
          description: Successfully retrieved tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndpointTag'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Tags
      summary: Create a new tag
      description: Create a new endpoint tag.
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - namespaceId
                - apiId
                - name
                - description
              properties:
                namespaceId:
                  type: string
                  example: "namespace-user"
                apiId:
                  type: string
                  example: "api-1"
                name:
                  type: string
                  example: "Users"
                description:
                  type: string
                  example: "User management endpoints"
      responses:
        '201':
          description: Tag created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointTag'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tags/{id}:
    get:
      tags:
        - Tags
      summary: Get tag by ID
      description: Retrieve a specific tag by its ID.
      operationId: getTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Tag ID
      responses:
        '200':
          description: Successfully retrieved tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointTag'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags:
        - Tags
      summary: Update tag
      description: Update an existing endpoint tag.
      operationId: updateTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Tag ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                apiId:
                  type: string
                  example: "api-1"
                name:
                  type: string
                  example: "UpdatedTagName"
                description:
                  type: string
                  example: "Updated description"
      responses:
        '200':
          description: Tag updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointTag'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Tags
      summary: Delete tag
      description: Delete an endpoint tag. Cannot delete if the tag is used by any endpoints.
      operationId: deleteTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Tag ID
      responses:
        '204':
          description: Tag deleted successfully
        '400':
          description: Cannot delete tag (used by endpoints)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Cannot delete tag: used by 5 endpoints"
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token from authentication flow
  schemas:
...
    EndpointTag:
      type: object
      required:
        - id
        - namespaceId
        - apiId
        - name
        - description
      properties:
        id:
          type: string
          description: Client-generated unique identifier
          example: "tag-1"
        namespaceId:
          type: string
          description: Namespace this tag belongs to
          example: "namespace-global"
        apiId:
          type: string
          description: Reference to the API this tag belongs to
          example: "api-1"
        name:
          type: string
          description: Tag name for OpenAPI spec
          example: "Users"
        description:
          type: string
          description: Tag description
          example: "User management endpoints"

    # ========================================================================
    # API Endpoint
    # ========================================================================
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: "Invalid API specification: endpoint path must start with /"
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing Clerk JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authenticated"